Convert thumbnail to png if webp
If youtube gives us a thumbnail in WEBP format, we have to convert it to PNG in order to embed it into the final video file.
We use Pillow for this as it is a very popular library. It might also come in handy later to convert other formats if needed.