Reduce console output
Overwrite messages for each segment download in the terminal output only, but keep writing the messages in the log files
The trick is to not use the console handler and print directly instead (to avoid adding the EOL character). We use a filter to do that by checking which function called and created the LogRecord.