Add on-download hook to spawn custom process
* Allows a user specified process and its arguments to be spawned
as a subprocess whenever we start downloading.
This is done only once, we are not responsible for the child process.
This should help mitigate failed recordings due to Youtube
throttling unexpectedly, since yt-dlp currently does not seem to
be affected by this issue.
* Fix missing output directory because exists() was not called
properly (the function pointer was tested instead... ugh).