Fix hook subprocess hanging
When logging for the subprocess was activated, the process was
stalled and did not complete its course, probably because of the logfile
being in a context handler.
We could solve this with a thread or something, but this would be overkill.
Let's simply fork and let the child process keep the file descriptors
from its parent should be enough to let it append to the logfile on its own.