Add --skip-download option
* This option is used to skip the download phase in livestream_saver
but to keep monitoring and reporting any active livestream detected.
This is useful when spawning yt-dlp as a subprocess and use only that to
download for example.
It could also be used to send e-mail notification whenever a livestream
has ended (but we did not download with livestream_saver, only with a
subprocess instead).
* Also fixed the e-mail notification option not being fetched as boolean
  from the config.