Fix warnings when detecting possible missing video Ids
When a new video is added to the channel, the last one in the previous
list gets moved to the second page (which we do not retrieve currently).
This led to repeated warning messages even though nothing was wrong.
Checking the number of video Ids alone might not be robust enough
anyway.
Since we are only interested in sending a warning if we suspect that we
just got logged out, only raise the warning if multiple video Ids go
missing suddenly; more than 3 should be enough as it is rare that
authors remove more than that in between scans.